This CL introduces an interface and implementing class to represent intervals of time and use them for both comparisons and scheduling.
It also includes an interface called ComparableElement which provides a general way for objects to be differentiated by objects (enums work particularly well).  Classes that implement this interface with a well-structured set of Comparisons  will be more efficient in Tree versions of containers.   

TimeIntervalEntry designates an object as being able to return a TimeInterval.

TimeIntervalScheduler shows how time intervals can be used to schedule an action on any object that implements TimeIntervalEntry.

Events can be scheduled for immediate action, future action at a specified time, or can be scheduled for action at at a randomly assigned moment of an interval.

The comparator provided for time intervals can be used to order a set  of scheduled events by adding them to TreeSet.





git-svn-id: svn://svn.caucho.com/home/svn/svnroot/resin/trunk@5215 9c94448d-38f1-0310-a231-d98308ff1ebf